some chnages

+++ support/reporter.ts
@@ -33,14 +33,16 @@ export class Reporter {
     public static cleanseDirectory(dir: string) {
         const fs = require('fs');
         const path = require('path');
-        fs.readdir(dir, (err, files) => {
-            if (err) throw err;
+        if (fs.existsSync(dir)) {
+            fs.readdir(dir, (err, files) => {
+                if (err) throw err;
 
-        for (const file of files) {
-        fs.unlink(path.join(dir, file), err => {
-            if (err) throw err;
+            for (const file of files) {
+            fs.unlink(path.join(dir, file), err => {
+                if (err) throw err;
             });
         }
         });
+    }
         }
 }
